refactor(admin): reports のDBアクセスを infrastructure へ切り出し
レポート一覧・種別集計クエリを infrastructure/report-repository.ts へ移設し、
interface は cacheLife 境界のみを担う。
