+++ apps/admin/src/features/reports/interface/queries.ts
@@ -1,26 +1,11 @@
-import { db } from '@repo/database';
-import { count, desc } from 'drizzle-orm';
 import { cacheLife } from 'next/cache';
+
+import { fetchReportsOverview } from '../infrastructure/report-repository';
 
 export const getReportsOverview = async () => {
   'use cache';
   cacheLife('minutes');
 
-  const [reports, typeCounts] = await Promise.all([
-    db.query.reportingReports.findMany({
-      orderBy: (reportTable) => [desc(reportTable.createdAt)],
-      limit: 100,
-    }),
-    db
-      .select({
-        type: db._schema.reportingReports.type,
-        count: count(),
-      })
-      .from(db._schema.reportingReports)
-      .groupBy(db._schema.reportingReports.type),
-  ]);
-
-  const totalCount = typeCounts.reduce((sum, t) => sum + t.count, 0);
-
-  return { reports, typeCounts, totalCount };
+  const overview = await fetchReportsOverview();
+  return overview;
 };
